replace compressor in `vortex-btrblocks`
