+++ vortex-btrblocks/benches/stats_calc.rs
@@ -10,7 +10,6 @@ mod benchmarks {
     use divan::Bencher;
     use vortex_array::arrays::PrimitiveArray;
     use vortex_array::validity::Validity;
-    use vortex_btrblocks::CompressorStats;
     use vortex_btrblocks::GenerateStatsOptions;
     use vortex_btrblocks::IntegerStats;
     use vortex_buffer::Buffer;
